+++ R/wrappers.R
@@ -136,3 +136,58 @@ with.reactibble <- function (data, expr, ...) {
 `$.reactibble` <- function(x, ...) {
   strip_reactive_col(.subset2(x, ...))
 }
+
+# This is necessary for binding functions
+# binding is only allowed between reactive columns of the same formula and
+# NAs (which are implicitly added when rowbinding)
+
+#' @export
+c.reactive_col <- function(...) {
+  dots <- list(...)
+  expr <- attr(..1, "reactible_col_def")
+  all_na <- all(sapply(dots[-1], is.na))
+  if(all_na) {
+    res <- NA
+    attr(res, "reactible_col_def") <- expr
+    return(res)
+  }
+  all_rc <- all(sapply(dots[-1], inherits, "reactive_col"))
+  if(!all_rc)
+    stop("Tried to bind a `reactive_col` with an incompatible data type")
+  unique_expr <- length(unique(lapply(dots, attr, "reactible_col_def"))) == 1
+  if(!unique_expr)
+    stop("Tried to bind `reactive_col` objects with incompatible column definitions")
+  res <- NextMethod()
+  attr(res, "reactible_col_def") <- expr
+  res
+}
+
+# This is necessary so dplyr::bind_rows reconstruct the reactibble and refreshes
+# it right
+
+#' @export
+dplyr_reconstruct.reactibble <- function (data, template)
+{
+  data   <- NextMethod()
+  data[] <- Map(function(x,y) {
+    attributes(x) <- y
+    x
+  }, data, lapply(template, attributes))
+  data
+}
+
+#' @export
+rbind.reactibble <- function(..., deparse.level = 1) {
+  data <- rbind.data.frame(..., deparse.level = 1)
+  # the main method does checks already so we do our checks
+  dots <- list(...)
+  rcs <- sapply(dots[[1]], inherits, "reactive_col")
+  nms <- names(which(rcs))
+  exprs1 <- sapply(.subset(dots[[1]], nms), attr, "reactible_col_def")
+  for(input in dots[-1]) {
+    exprs <- sapply(.subset(input, nms), attr, "reactible_col_def")
+    if(!identical(exprs, exprs1))
+      stop("Tried to bind a `reactive_col` to an incompatible object.")
+  }
+  refresh(data)
+}