+++ R/zz.R
@@ -8,6 +8,7 @@
   registerS3method("anti_join", "reactibble", left_join.reactibble, asNamespace("dplyr"))
   registerS3method("semi_join", "reactibble", left_join.reactibble, asNamespace("dplyr"))
   registerS3method("nest_join", "reactibble", left_join.reactibble, asNamespace("dplyr"))
+  registerS3method("dplyr_reconstruct", "reactibble", dplyr_reconstruct.reactibble, asNamespace("dplyr"))
   #registerS3method("as_tibble", "reactibble", as_tibble.reactibble, asNamespace("tibble"))
   registerS3method("vec_ptype_abbr", "reactive_col", vec_ptype_abbr.reactive_col, asNamespace("vctrs"))
   registerS3method("pillar_shaft", "reactive_col", pillar_shaft.reactive_col, asNamespace("pillar"))