+++ NEWS.md
@@ -20,5 +20,7 @@
 * When printing a tibble and the "reactibble.autorefresh" option is `TRUE` mark
   columns as "unsynced" colored in red.
 * Optimize process_reactive_dots so reactive columns are coomputed only in the end
+* Use quosures instead of expressions as column definitions
+* implement temporary (brittle) solution to rbind and bind_rows